Add Elastic namespace and missing controller actions
