+++ app/services/elastic/posts/create_service.rb
@@ -1,0 +1,20 @@
+module Elastic
+  module Posts
+    class CreateService
+      def initialize(post_params)
+        @params = post_params
+      end
+
+      def call
+        post = Post.create!(params)
+        ElasticIndexerJob.perform_later(:create, post)
+
+        post
+      end
+
+      private
+
+      attr_reader :params
+    end
+  end
+end
