--- app/services/posts/create_service.rb
@@ -1,16 +1,0 @@
-module Posts
-  class CreateService
-    def initialize(post_params)
-      @params = post_params
-    end
-
-    def call
-      post = Post.create!(params)
-      ElasticIndexerJob.perform_later(:create, post)
-    end
-
-    private
-
-    attr_reader :params
-  end
-end
