+++ app/views/elastic/posts/_post.html.erb
