+++ app/views/elastic/posts/edit.html.erb
@@ -1,0 +1,10 @@
+<h1>Editing post</h1>
+
+<%= render "form", post: @post %>
+
+<br>
+
+<div>
+  <%= link_to "Show this post", [:elastic, @post] %> |
+  <%= link_to "Back to posts", elastic_posts_path %>
+</div>
