+++ app/views/elastic/posts/index.html.erb
@@ -6,9 +6,9 @@
   <% @posts.each do |post| %>
     <%= render post %>
     <p>
-      <%= link_to "Show this post", post %>
+      <%= link_to "Show this post", [:elastic, post] %>
     </p>
   <% end %>
 </div>
 
-<%= link_to "New post", new_post_path %>
+<%= link_to "New post", new_elastic_post_path %>
