+++ app/views/elastic/posts/show.html.erb
@@ -1,0 +1,10 @@
+<p style="color: green"><%= notice %></p>
+
+<%= render @post %>
+
+<div>
+  <%= link_to "Edit this post", edit_elastic_post_path(@post) %> |
+  <%= link_to "Back to posts", elastic_posts_path %>
+
+  <%= button_to "Destroy this post", elastic_post_path(@post), method: :delete %>
+</div>
